% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ppet.R
\name{add_ppet}
\alias{add_ppet}
\title{Add P-PET Column to Data Frame}
\usage{
add_ppet(data, precip_col, pet_col, new_col = "P_PET")
}
\arguments{
\item{data}{A data frame.}

\item{precip_col}{Character. Name of precipitation column.}

\item{pet_col}{Character. Name of PET column.}

\item{new_col}{Character. Name for new P-PET column. Default "P_PET".}
}
\value{
Data frame with P-PET column added.
}
\description{
Add P-PET Column to Data Frame
}
