% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rswc.R
\name{add_rswc}
\alias{add_rswc}
\title{Add RSWC Column to Data Frame}
\usage{
add_rswc(data, swc_col, new_col = "RSWC", ...)
}
\arguments{
\item{data}{A data frame.}

\item{swc_col}{Character. Name of the SWC column.}

\item{new_col}{Character. Name for the new RSWC column. Default "RSWC".}

\item{...}{Additional arguments passed to calc_rswc().}
}
\value{
Data frame with RSWC column added.
}
\description{
Convenience function to add RSWC as a new column to a data frame.
}
\examples{
\dontrun{
data <- add_rswc(data, swc_col = "SoilWaterContent")
}

}
