% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_json.R
\name{export_data_json}
\alias{export_data_json}
\title{Export Data Frame to JSON}
\usage{
export_data_json(data, path, pretty = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{path}{Character. Output file path.}

\item{pretty}{Logical. Pretty-print? Default TRUE.}
}
\value{
Path to written file (invisibly).
}
\description{
Simple export of a data frame to JSON, useful for passing data
to the thermal-vegetation-monitoring web interface.
}
