% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_treatment.R
\name{print.grassland_analysis}
\alias{print.grassland_analysis}
\title{Print Method for grassland_analysis}
\usage{
\method{print}{grassland_analysis}(x, ...)
}
\arguments{
\item{x}{A grassland_analysis object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for grassland_analysis
}
