% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_urep.R
\name{print.urep_data}
\alias{print.urep_data}
\title{Print Method for urep_data}
\usage{
\method{print}{urep_data}(x, ...)
}
\arguments{
\item{x}{A urep_data object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for urep_data
}
